FINAL
Removed save Picture feature, because done testing and that'd be kinda weird to keep it in.  Added settings button so you can manually adjust threshold and distance width, more so for testing, but could be handy for threshold and lighting situations. Cleaned up query and how it searches google, fixed errors in returning foot length instead of equivalent shoe Size. Removed a majority of the Width functionality, because it's unnecessary and can just use the Length methods since you're turning your foot anyways. Updated Pixel measures to reflect Length and Width, and lastly changed the title of Webpage to show the estimated foot size, should you get a link that's not a direct product page