Not final yet, expect at least one more commit
Bothered me that my attempt to open them fullscreen was just setting the window to the dimensins of the screen, but it wasn't truly maximized. Googled it and found a way to make it truly maximized, so i went in and changed it. Also deleted all console prints, as ideally this will be run as a jar and thus won't have a console to print.